+++ lib/langtool_web.ex
@@ -23,6 +23,8 @@ defmodule LangtoolWeb do
       import Plug.Conn
       import LangtoolWeb.Router.Helpers
       import LangtoolWeb.Gettext
+      import LangtoolWeb.Errors
+      use LangtoolWeb.Auth
     end
   end
 
@@ -32,7 +34,10 @@ defmodule LangtoolWeb do
                         namespace: LangtoolWeb
 
       # Import convenience functions from controllers
-      import Phoenix.Controller, only: [get_csrf_token: 0, get_flash: 2, view_module: 1]
+      import Phoenix.Controller, only: [get_csrf_token: 0, get_flash: 1, get_flash: 2, view_module: 1]
+
+      # Helper for checking singning
+      import LangtoolWeb.Helpers.Auth, only: [signed_in?: 1]
 
       # Use all HTML functionality (forms, tags, etc)
       use Phoenix.HTML