+++ lib/langtool_web/views/dashboard_view.ex
@@ -1,0 +1,3 @@
+defmodule LangtoolWeb.DashboardView do
+  use LangtoolWeb, :view
+end