+++ lib/langtool_web/views/email_view.ex
@@ -1,0 +1,3 @@
+defmodule LangtoolWeb.EmailView do
+  use LangtoolWeb, :view
+end