+++ lib/langtool_web/views/user_view.ex
@@ -1,0 +1,3 @@
+defmodule LangtoolWeb.UserView do
+  use LangtoolWeb, :view
+end