+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2019-01-05
+### Added
+- user creation by signup
+- user mailer with confirmation email
+- confirmation process
+- alerts block
+- signout process
+- signin process
+- presenter for validation errors
+- dashboard page with auth check for access
+- authorization policies system with dashboard policy
+
 ## [0.1.0] - 2019-01-01
 ### Added
 - background job for task file handling